chore(bulkimport): increase robustness for final database writes
Since we're importing we know beforehand how many database flushes will happen, hence we know how many times the
  case b := <-i.Processor.WritebackChan():
in the processor's select will happen: exactly len(keys) times.

To make sure we wait until this is true, we wait forever until dbWritebackCallAmt is equal to len(keys).
To further strengthen our logic here we only increment dbWritebackCallAmt when the WritebackChan acquires a lock
on dbMutex, so that when this infinite for cycle will actually break, we block again until the database func call has
finished writing.
After that, we acquire the lock and continue with our own way.