fix: ignore iris strange trace (#127)
insert balances fails because the amount field has non-UTF-8 characters.

This is due to a strange iris trace that is interpreted as bank balance trace with a value `\n$\n\x05uiris\x12\x1b509625143506063788050678455`. When this value is unmarshaled into a `sdk.Coin`, it doesn't trigger any error but fills the `sdk.Coin.Amount` field with invalid UTF-8 characters, leading to the database error mentioned earlier.

Actually, this iris trace is not a balance bank trace, because the format of the key is `0x02+len(addr)+addr`, whereas the balance key format is `0x02+len(addr)+addr+denom`.

So, despite we didn't manage to find the module corresponding to this trace, what appears to be the best fix is to ensure the bank processor ignores this trace.

- This commit adds tests to the existing code: edef1fcfb9ce629723dd982b337dff8cb12279a4
- This commit ensures the bank processor ignores traces like the one of Iris: 8c0c7df41ff4d1f6dee81318ddbcab7122dc0992